Consolidate and use OnlyGeoTestBase
