+++ test/celeritas/ext/Vecgeom.test.cc
@@ -12,8 +12,10 @@
 #include "corecel/io/Repr.hh"
 #include "corecel/math/NumericLimits.hh"
 #include "corecel/sys/Device.hh"
+#include "celeritas/GeantTestBase.hh"
 #include "celeritas/GlobalGeoTestBase.hh"
 #include "celeritas/GlobalTestBase.hh"
+#include "celeritas/OnlyGeoTestBase.hh"
 #include "celeritas/ext/LoadGdml.hh"
 #include "celeritas/ext/VecgeomData.hh"
 #include "celeritas/ext/VecgeomParams.hh"
@@ -35,16 +37,10 @@ namespace test
 #    define TEST_IF_CELERITAS_CUDA(name) DISABLED_##name
 #endif
 
-// Always compile and sometimes disable tests that require Geant4
-#if CELERITAS_USE_GEANT4
-#    define TEST_IF_CELERITAS_GEANT(name) name
-#else
-#    define TEST_IF_CELERITAS_GEANT(name) DISABLED_##name
-#endif
-
 //---------------------------------------------------------------------------//
 // TESTS
 //---------------------------------------------------------------------------//
+
 class VecgeomTestBase : virtual public GlobalTestBase
 {
   public:
@@ -77,18 +73,6 @@ class VecgeomTestBase : virtual public GlobalTestBase
 
     //! Find linear segments until outside
     TrackingResult track(const Real3& pos, const Real3& dir);
-
-  protected:
-    SPConstParticle    build_particle() final { CELER_ASSERT_UNREACHABLE(); }
-    SPConstCutoff      build_cutoff() final { CELER_ASSERT_UNREACHABLE(); }
-    SPConstPhysics     build_physics() final { CELER_ASSERT_UNREACHABLE(); }
-    SPConstTrackInit   build_init() override { CELER_ASSERT_UNREACHABLE(); }
-    SPConstAction      build_along_step() final { CELER_ASSERT_UNREACHABLE(); }
-    SPConstMaterial    build_material() final { CELER_ASSERT_UNREACHABLE(); }
-    SPConstGeoMaterial build_geomaterial() final
-    {
-        CELER_ASSERT_UNREACHABLE();
-    }
 
   private:
     HostStateStore host_state;
@@ -151,7 +135,9 @@ void VecgeomTestBase::TrackingResult::print_expected()
 
 //---------------------------------------------------------------------------//
 
-class FourLevelsTest : public VecgeomTestBase, public GlobalGeoTestBase
+class FourLevelsTest : public VecgeomTestBase,
+                       public GlobalGeoTestBase,
+                       public OnlyGeoTestBase
 {
   public:
     const char* geometry_basename() const final { return "four-levels"; }
@@ -428,32 +414,29 @@ TEST_F(FourLevelsTest, TEST_IF_CELERITAS_CUDA(device))
 // CONSTRUCT FROM GEANT4 (TODO)
 //---------------------------------------------------------------------------//
 
-#define GeantBuilderTest TEST_IF_CELERITAS_GEANT(GeantBuilderTest)
-class GeantBuilderTest : public VecgeomTestBase, virtual public GlobalTestBase
+class GeantBuilderTestBase : public VecgeomTestBase,
+                             virtual public GeantTestBase
 {
   public:
     static void SetUpTestCase()
     {
         // Make sure existing VecGeom geometry has been cleared
-        test::GlobalGeoTestBase::reset_geometry();
-    }
-
-    void SetUp() override
-    {
-        VecgeomTestBase::SetUp();
-        world_volume_
-            = load_gdml(this->test_data_path("celeritas", "four-levels.gdml"));
+        GlobalGeoTestBase::reset_geometry();
     }
 
     SPConstGeo build_geometry() override
     {
-        CELER_NOT_IMPLEMENTED("build_geometry");
-        // return std::make_shared<VecgeomParams>(world_volume.get());
+        // TODO:
+        return nullptr;
+        // return std::make_shared<VecgeomParams>(this->get_world_volume());
     }
 
-  private:
-    UPG4PhysicalVolume world_volume_;
+    bool      enable_fluctuation() const final { return false; }
+    bool      enable_msc() const final { return false; }
+    bool      combined_brems() const final { return false; }
+    real_type secondary_stack_factor() const final { return 0; }
 };
+
 //---------------------------------------------------------------------------//
 } // namespace test
 } // namespace celeritas
